Use wiremock for SQS testing.
This will allows us to check if an API call to SQS has been made which
could be useful for testing the change visibility timeout changes.

Also, this may be better because we're not relying on any extra third
party libraries as we were already using WireMock for KMS calls.