Create and delete queue with each test.
Originally we were receiving all messages before the tests and deleting
them but because of visibility timeouts, we weren't always getting all
of the messages, so we couldn't delete them, so they persisted between
tests.
The easiest and cleanest way around it is to delete and recreate the
queue for each test.