Change the way we create the event.
Each call to sqsClient.receive changes the receipt handle and only the most recent receipt handle is checked when deleting a message. This mirrors sqs behaviour where only the most recent recipient can delete the message. What was happening here when there were two messages being sent to the input queue was:

* Message 1 sent to the input queue
* sqsClient.receive called which changes the reciept handle for message 1 and returns the new receipt handle
* Receipt handle added to record for message 1
* Message 2 is sent to the input queue
* sqsClient.receive called which changes the receipt handle for message 1 *and* Message 2 and returns both new handles
* Receipt handle for message 2 is added to the record.
* The receipt handle for message 1 which is set in the record, is now wrong as it has been replaced by the second call to sqsClient.receive.
* When you try to call delete on message 1, it fails (silently) because you're trying to use an out of date receipt handle.

The fix is to make sure that sqsClient.receive is only called once in the test setup even if there are two messages.