Set visibility timeout on error.
This uses the IO method handleErrorWith. It relies on having valid
decoded json so it can use the receipt handle to change the message
visibility so if there is a failure decoding the json, the visibility
timeout isn't changed but there's not much we can do about that.

I've added a couple of tests to check that the two error types set the
timeout correctly. They're technically being tested further up but I've
added them in so the description states explicitly what's happening.